pydiag: address comments in PR
* Change copyright year
* Add copyright header on files missing it
* Rename pydiag to pyfpgadiag
* Add pytlint directive to ignore E0611 when importing opae.fpga
