+++ tools/CMakeLists.txt
@@ -36,7 +36,7 @@ add_subdirectory(extra/coreidle)
 add_subdirectory(extra/userclk)
 add_subdirectory(extra/packager)
 add_subdirectory(extra/c++utils)
-add_subdirectory(extra/pydiag)
+add_subdirectory(extra/pyfpgadiag)
 add_subdirectory(extra/libopae++)
 add_subdirectory(extra/mmlink)
 add_subdirectory(extra/ras)
